Refactoring as package with easier testing
