+++ idg2sl/__init__.py
@@ -1,0 +1,6 @@
+from .synthetic_lethal_interaction import SyntheticLethalInteraction
+from .entrez_parser import EntrezParser
+from .idg2sl_parser_human import *
+
+
+__all__ = ["SyntheticLethalInteraction", "EntrezParser"]
