+++ test/data/luo_small.tsv
@@ -1,0 +1,11 @@
+Symbol	Accession	v2SH	Sequence	Mean.DLD1	SD.DLD1	Mean.HCT116	SD.HCT116
+ANAPC1	NM_022662	v2HS_98169	AATAGACATATCCATGGTGCCT	82.2	3.2	48.3	3.7
+ANAPC1	NM_022662	v2HS_98171	TTTATGCAAACAGTTAAATGCT	55.1	1.9	81	6.9
+ANAPC1	NM_022662	v2HS_98167	TATTAGTAATCGTTTCTGTCCA	75.5	1.7	104.9	10.7
+ANAPC4	NM_013367	v2HS_254660	ATACAATGGAATACAGATTGCT	92.1	5.8	51.4	1.7
+ANAPC4	NM_013367	v2HS_64633	TATACTAAAGACAAAGGCAGCT	84.8	4.4	59.2	4.8
+BXDC1	NM_032194	v2HS_257526	ATTACTAGATTATTTGGCCGCT	84.6	4.6	89.3	3
+BXDC2	NM_018321	v2HS_275380	TTGAACAAGGAATTTAGCAGAT	79.4	7.9	64.7	2.7
+BXDC5	NM_025065	v2HS_235217	TTTACCAAAGTTCAAGATAGGC	71.4	3.9	67.1	3.2
+C19orf2	NM_134447	v2HS_23231	TTATAAGACAATTTATCAGGCA	78.2	2.9	67.1	4.5
+C20orf121	NM_024331	v2HS_17999	TGTAGATTTCATCCAGTCTCTC	85.2	2.9	83.1	2
